requirements.txt
langchain
langchain-google-genai
langchain-groq
langgraph
pypdf
python-docx
python-dotenv
